# soulsdle-api-spec.yaml
openapi: 3.0.2
info:
  version: 1.0.0
  title: API Soulsdle - Jeu de devinettes sur les boss FromSoftware
  description: API REST pour un jeu de type Wordle avec les boss des jeux FromSoftware (Bloodborne, Sekiro, Dark Souls III)
  contact:
    name: "Cyrian Torrejon | Sinfix"
    email: cyrian.torrejon@etu.univ-grenoble-alpes.fr
servers:
  - url: http://localhost:3000
    description: Serveur de développement local
paths:
...
  /joueurs:
    get:
      tags:
        - Joueurs
      summary: Liste tous les joueurs
      parameters:
        - $ref: "#/components/parameters/pageParam"
        - $ref: "#/components/parameters/limitParam"
      responses:
        "200":
          description: Liste des joueurs
          content:
            application/json:
              schema:
                type: object
                properties:
                  joueurs:
                    type: array
                    items:
                      $ref: "#/components/schemas/Joueur"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
    post:
      tags:
        - Joueurs
      summary: Crée un nouveau joueur
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Joueur"
      responses:
        "201":
          description: Joueur créé avec succès
  /joueurs/{id}:
    get:
      tags:
        - Joueurs
      summary: Détails d'un joueur spécifique
      parameters:
        - name: id
          in: path
          description: ID du joueur
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Informations détaillées du joueur
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Joueur"
        "404":
          description: Joueur non trouvé
    put:
      tags:
        - Joueurs
      summary: Met à jour un joueur
      parameters:
        - name: id
          in: path
          description: ID du joueur
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Joueur"
      responses:
        "200":
          description: Joueur mis à jour
        "404":
          description: Joueur non trouvé
    delete:
      tags:
        - Joueurs
      summary: Supprime un joueur
      parameters:
        - name: id
          in: path
          description: ID du joueur
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Joueur supprimé
        "404":
          description: Joueur non trouvé
...
components:
  schemas:
...
    Joueur:
      type: object
      properties:
        id:
          type: string
          description: Identifiant unique du joueur
          example: "joueur-001"
        pseudo:
          type: string
          description: Pseudonyme du joueur
          example: "SoulsHunter"
        email:
          type: string
          format: email
          description: Adresse email du joueur
          example: "soulshunter@example.com"
        partiesGagnees:
          type: integer
          description: Nombre de parties gagnées
          example: 12
        streakActuelle:
          type: integer
          description: Série actuelle de victoires consécutives
          example: 5
        meilleureStreak:
          type: integer
          description: Meilleure série de victoires consécutives
          example: 12
      required:
        - id
        - pseudo
        - email
        - partiesGagnees
        - streakActuelle
        - meilleureStreak
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
  parameters:
    pageParam:
      name: page
      in: query
      description: Numéro de page (commence à 1)
      required: false
      schema:
        minimum: 1
        type: integer
        format: int32
        default: 1
    limitParam:
      name: limit
      in: query
      description: Nombre maximum d'éléments par page
      required: false
      schema:
        maximum: 100
        minimum: 1
        type: integer
        format: int32
        default: 20